feat(parity): custom-spike-author -- 7/9 remaining bundled routines promoted
Corrects the "custom_spike" bucket scope: the ~103 figure from the
original acceleration audit was misattributed (an unrelated, unmerged
field-module cascade benchmark, never dispatched into ff4-gnw). The real,
registry-tracked bucket was 11 routines; this closes 7 of them plus the 2
already promoted earlier this session (DrawMonsterSprite_c,
BuildOAMEntries_c), using the same extract-to-standalone-file +
region-compare-spike pattern proven on 2026-06-30.

Promoted L1->L2: Mult8_btlgfx_c, HardMult_btlgfx_c, IncrTextPtr_c,
CheckSpriteVisible_c, InitMonsterAnim_c (all bundled in
ff4-gnw/battle/{btlgfx_prim,btlgfx_monsters}.c).

Correctly LEFT AT L1, with evidence explaining why (not silently dropped):
- BackAttackYOffset_s_c/_l_c: generate_spike.py has no register-output
  comparison mode -- their only observable effect is the accumulator, so
  the generated spike compares two hardcoded zeros (always "passes",
  proves nothing). Confirmed not isolated to these two (spike_getantielem_*,
  spike_max99_* hit the same hollow branch). See MemPalace wing=ff4-gnw
  room=obstacles-and-solutions and the updated
  .claude/skills/ff4-custom-spike-author/SKILL.md.
- UpdateMonsterAnim_c: genuinely incomplete C port, not a spike gap -- its
  "zero-path" (re-init branch) depends on three un-ported sub-routines
  and a stale cpu->y bug. The whole-game oracle never caught it because
  the tested savestates don't exercise that branch. Real porting work,
  out of scope for this session.
- UpdateBG2ScrollSkip_c: NOT a routine bug at all -- the disassembly
  (upstream/notes/ff4j-sfc.asm) has a 2-byte address offset error at this
  location. The real ROM's $00:F535 entry point loads its own guard byte
  ($C9), but the currently-dispatched UpdateBG2ScrollSkip_c assumes the
  caller pre-loaded it into A. Confirmed by direct ROM-byte inspection and
  an independently re-verified fix (0/2000 fails with the corrected guard
  source vs. failures at scale with the current one). This is a live
  dispatch bug in ff4-gnw/dispatch_all.c, not touched here pending
  confirmation on the fix approach -- see MemPalace for full detail.

Also fixes a SECOND, independent 8-bit-truncation-before-widening bug in
BuildOAMEntries_c (slot_w), found by re-running its spike at a higher
trial count than the one used for its original promotion -- the same bug
class as the first fix (y_row), on a different line. Folded into the
skill: always re-verify at a higher trial count before trusting a "clean"
spike; a ~0.1-1% divergence rate can pass a 300-trial run by chance.

parity/Makefile gains the auto-generated build targets for the 9 new
standalone spikes (generate_spike.py's own convention, not hand-edited).

+++ translator/port/battle/BuildOAMEntries.c
@@ -31,7 +31,10 @@ void BackAttackYOffset_s_c(Snes *snes) {
 
 void BuildOAMEntries_c(Snes *snes) {
     uint8_t slot = (uint8_t)(snes->cpu->x);
-    uint16_t y_row = (uint16_t)((uint16_t)slot << 2);
+    /* TXA/ASL/ASL/TAY, A is 8-bit (mf=true): each ASL truncates to 8 bits,
+     * so the real hardware wraps at 256, not 65536 — mask before widening
+     * or slot>=64 diverges from the original. */
+    uint16_t y_row = (uint16_t)(uint8_t)(slot << 2);
     uint8_t tile_count = snes->cart->rom[LOROM(0x16, 0xFD15) + slot];
     uint16_t loop_limit = (uint16_t)tile_count << 2;
     snes->ram[0x0E] = (uint8_t)(loop_limit & 0xFF);
@@ -52,7 +55,11 @@ void BuildOAMEntries_c(Snes *snes) {
         snes->ram[0xF07D + y_row]++;
     }
 
-    uint16_t slot_w = (uint16_t)((uint16_t)slot << 1);
+    /* TXA/ASL/TAX ($02:DD26-28): A is 8-bit (mf=true), the ASL truncates
+     * to 8 bits before TAX widens it back to the 16-bit X -- truncate
+     * before widening or slot>=128 silently diverges (same bug class as
+     * y_row above, found by a higher-trial-count re-run, 2026-07-05). */
+    uint16_t slot_w = (uint16_t)(uint8_t)(slot << 1);
     uint8_t x_base = snes->ram[0x6CF3 + slot_w];
     snes->ram[0x12] = x_base;
     uint8_t y_base = snes->ram[0x6CF4 + slot_w];
